Move function to get user created date for Zendesk tickets
We need to use this in more places now, so it makes sense to move it to
a more general file.
