+++ app/main/views/feedback.py
@@ -16,6 +16,7 @@ from app.models.feedback import (
     QUESTION_TICKET_TYPE,
 )
 from app.utils import hide_from_search_engines
+from app.utils.user import get_user_created_at_for_ticket
 
 bank_holidays = BankHolidays(use_cached_holidays=True)
 
@@ -249,10 +250,3 @@ def get_zendesk_ticket_type(ticket_type):
         return NotifySupportTicket.TYPE_INCIDENT
 
     return NotifySupportTicket.TYPE_QUESTION
-
-
-def get_user_created_at_for_ticket(user) -> datetime | None:
-    if user.is_authenticated:
-        return datetime.strptime(user.created_at, "%Y-%m-%dT%H:%M:%S.%fZ").replace(tzinfo=pytz.utc)
-
-    return None
